requirements.txt
absl-py==2.1.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asciitree==0.3.3
asttokens==2.2.1
attrs==22.2.0
audioread==3.0.1
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
brotlipy==0.7.0
certifi==2021.10.8
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
comm==0.1.2
conda==4.11.0
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262147379/work
contourpy==1.0.7
cryptography @ file:///tmp/build/80754af9/cryptography_1639414572950/work
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
et-xmlfile==1.1.0
executing==1.2.0
fasteners==0.19
fastjsonschema==2.16.3
ffmpeg-python==0.2.0
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.39.0
fqdn==1.5.1
fsspec==2024.9.0
future==1.0.0
gpytorch==1.13
grpcio==1.66.2
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jams==0.3.4
jaxtyping==0.2.19
jedi==0.18.2
Jinja2==3.1.2
joblib==1.4.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.4.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lazy_loader==0.4
librosa==0.10.2.post1
linear-operator==0.5.3
llvmlite==0.43.0
Markdown==3.7
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mir_eval==0.7
mistune==2.0.5
mpmath==1.3.0
msgpack==1.1.0
musdb==0.4.2
museval==0.4.1
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.2.1
norbert==0.2.1
notebook==6.5.3
notebook_shim==0.2.2
numba==0.60.0
numcodecs==0.12.1
numpy==2.0.2
nussl==1.1.9
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
pooch==1.8.2
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==5.28.2
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==24.9.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.14.0
pyloudnorm==0.1.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-ignite==0.5.1
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.1
qtconsole==5.4.1
QtPy==2.3.0
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016711199/work
scikit-learn==1.5.2
scipy==1.13.1
Send2Trash==1.8.0
simplejson==3.19.3
six @ file:///tmp/build/80754af9/six_1623709665295/work
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4
sox==1.5.0
soxbindings==1.2.3
soxr==0.5.0.post1
stack-data==0.6.2
stempeg==0.2.3
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
terminado==0.17.1
termtables==0.2.4
threadpoolctl==3.5.0
tinycss2==1.2.1
torch==2.4.1
tornado==6.2
tqdm @ file:///tmp/build/80754af9/tqdm_1635330843403/work
traitlets==5.9.0
triton==3.0.0
typeguard==4.3.0
typing_extensions==4.12.2
uri-template==1.2.0
urllib3==1.26.7
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==3.0.4
widgetsnbextension==4.0.5
zarr==2.18.2
zipp==3.15.0
